requirements.txt
airtable-python-wrapper==0.15.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
numpy==1.19.2
packaging==20.4
pyparsing==2.4.7
regex==2020.10.15
requests==2.24.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
tokenizers==0.8.1rc2
tqdm==4.50.2
transformers==3.3.1
urllib3==1.25.10
Werkzeug==1.0.1
